bug fixed, db type data

+++ database/migrations/2022_04_14_055739_create_detail_produk_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->bigInteger("idProduk");
             $table->bigInteger("idKategori");
-            $table->integer("jumlahStok");
+            $table->smallInteger("jumlahStok");
             $table->timestamps();
         });
     }
